Разбил код на модули, настроил webpack и eslint
