+++ src/components/daylist.js
@@ -1,0 +1,3 @@
+const createDayListTemplate = () => `<ul class="trip-days"></ul>`;
+
+export {createDayListTemplate};
